// libs/domain/src/user-management/enum/permission-name.enum.ts
export enum PermissionNames {
  baseInfo = 'BaseInfo',
  personCompany = 'PersonCompany',
  inventoryType = 'InventoryType',
  inventory = 'Inventory',
  fiscalPeriod = 'FiscalPeriod',
  fiscalPeriodStatus = 'FiscalPeriodStatus',
  goodsOrderPoint = 'GoodsOrderPoint',
  delivery = 'Delivery',
  receipt = 'Receipt',
  transfer = 'Transfer',
  inventoryCounting = 'InventoryCounting',
  inventoryAccounting = 'InventoryAccounting',
  goodsServiceOrder = 'GoodsServiceOrder',
  priceEstimate = 'PriceEstimate',
  purchase = 'Purchase',
  purchaseReturn = 'PurchaseReturn',
  PurchaseReference = 'PurchaseReference',
  user = 'User',
  role = 'Role',
  permission = 'Permission',
  goodsServiceCategory = 'GoodsServiceCategory',
  goodsServiceProperty = 'GoodsServiceProperty',
  measurement = 'Measurement',
  goodsService = 'GoodsService',
  goodsServiceHeading = 'GoodsServiceHeading',
  organization = 'Organization'
}